+++ thoth-api/migrations/v0.12.3/down.sql
@@ -3,3 +3,10 @@ ALTER TABLE series
 
 ALTER TABLE series
     ALTER COLUMN issn_digital SET NOT NULL;
+
+ALTER TABLE work
+    DROP CONSTRAINT work_active_withdrawn_date_check,
+    DROP CONSTRAINT work_inactive_no_withdrawn_date_check,
+    DROP CONSTRAINT work_withdrawn_date_after_publication_date_check,
+    DROP COLUMN withdrawn_date;
+